Cleaned the Code and Made Updates.
No Errors during the Process of Login and Logoff. No testing was done for registration.

Note that the User's passwords are NOT encrypted as the note was made, it is Hashed, Also they are UNSALTED HASHES so easy to crack with John the Ripper.

The updates to the code has made it not have any errors or breaks in the code that can slow it down. More mods can be made to make it better and more secure (like using mysqli through out etc.) but I have done what I can for a bit.